The background scripts now receive a JSON object instead of a javascript array.  Subsequently, I removed your previous code that "stringified" the translations.  I am assuming that since the translations are now made of an element already placed into JSON object format that it would not require encoding it as JSON.  I may be wrong--interested to hear your input.
I also changed the variable that is to be passed back to the content scripts.  However, I am having trouble debugging the variable I created.  The variable I am using is "merged." The intent of the variable merged is to join the original_text and translation JSON objects into one JSON object where the original word and the translations will be in a key/value format.